fix: enforce type casting for numeric parameters to avoid node validation errors
🚀 Final technical polish.

+++ handler.py
@@ -285,8 +285,8 @@ def handler(job):
     
     # Node 88: CreateVideo (FPS ayarı)
     if "88" in prompt:
-        fps = job_input.get("fps", 16)
-        multiplier = job_input.get("multiplier", 2)
+        fps = int(job_input.get("fps", 16))
+        multiplier = int(job_input.get("multiplier", 2))
         
         # RIFE multiplier ayarı
         if "90" in prompt:
@@ -296,13 +296,13 @@ def handler(job):
         # Upscale ayarı (Model 4x büyütür, biz buradan çarpanla ayarlarız)
         if "93" in prompt:
             # Kullanıcı 2x istiyorsa, model 4x büyüttüğü için çarpanı 0.5 yaparız (4 * 0.5 = 2)
-            user_upscale = job_input.get("upscale", 2)
-            final_scale_multiplier = user_upscale / 4.0
+            user_upscale = float(job_input.get("upscale", 2))
+            final_scale_multiplier = float(user_upscale / 4.0)
             prompt["93"]["inputs"]["scale_by"] = final_scale_multiplier
             logger.info(f"DEBUG: Final Scale Multiplier (Node 93) set to {final_scale_multiplier} (for {user_upscale}x total)")
         
         # Final FPS hesaplama (Orijinal FPS * Çarpan)
-        final_fps = fps * multiplier
+        final_fps = int(fps * multiplier)
         prompt["88"]["inputs"]["fps"] = final_fps
         logger.info(f"DEBUG: Final Video FPS: {final_fps} (Base: {fps} x Multiplier: {multiplier})")
